refactor: style 파일 분리안된 컴포넌트 수정
