+++ src/components/svg/CloseIcon/CloseIcon.styles.js
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled';
+
+const SVG = styled.svg`
+  transform: scale(${props => props.scale});
+`;
+
+export { SVG };
