+++ src/components/svg/ShoppingCartIcon/ShoppingCartIcon.styles.js
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+const SVG = styled.svg`
+  cursor: pointer;
+  transform: ${({ scale }) => scale && `scale(${scale})`};
+`;
+
+export { SVG };
