+++ src/components/svg/TrashIcon/TrashIcon.js
@@ -1,10 +1,5 @@
 import PropTypes from 'prop-types';
-import styled from '@emotion/styled';
-
-const SVG = styled.svg`
-  cursor: pointer;
-  transform: ${({ scale }) => scale && `scale(${scale})`};
-`;
+import { SVG } from './TrashIcon.styles';
 
 const TrashCanIcon = ({ scale }) => (
   <SVG scale={scale} xmlns="http://www.w3.org/2000/svg" width="22" height="22" viewBox="0 0 22 22" fill="none">
