+++ src/components/templates/GlobalNavbar/GlobalNavbar.js
@@ -1,19 +1,7 @@
-import styled from '@emotion/styled';
 import { Link } from 'react-router-dom';
 import { NavBar, ShoppingCartIcon } from '../..';
 import { ROUTE } from '../../../constants';
-
-const LogoContainer = styled(Link)`
-  font-size: 2rem;
-  display: flex;
-  align-items: center;
-  color: white;
-  font-weight: 900;
-
-  & > *:not(:last-child) {
-    margin-right: 10px;
-  }
-`;
+import { LogoContainer } from './GlobalNavBar.styles';
 
 const GlobalNavbar = () => {
   const Logo = (
